refactor: move step components to Checkout/components
