+++ src/features/Checkout/components/BagStep/style.scss
