+++ src/features/Checkout/components/SuccessStep/SuccessStep.tsx
@@ -1,18 +1,27 @@
-import { BagContent } from "../components/BagContent"
-import { Billing, Product } from "../components/BagContent/types"
-import { BillingSummary } from "../components/BillingSummary"
+import { BagContent } from "../BagContent"
+import { Billing, Product } from "../BagContent/types"
+import { BillingSummary } from "../BillingSummary"
+import { SuccessfulCard } from "../SuccessfulCard"
 import './style.scss'
 
 interface BagStepProps { 
+  payload: any
   products: Product[]
   billing: Billing,
   goToNextHandler: () => void
 }
 
-export function BagStep({ products, billing, goToNextHandler }: BagStepProps) {
+export function SuccessStep({ payload, products, billing, goToNextHandler }: BagStepProps) {
   return (
     <div className='bag-step'>
-      <BagContent key="Sacola" products={products} />
+      <div className="scrollable-content">
+        <SuccessfulCard payload={payload}  />
+        <BagContent
+          key="Sacola" 
+          title='Produtos' 
+          products={products}
+          showPrices={false} />
+      </div>
       <section className='summary-action'>
         <BillingSummary 
           shippingTotal={billing.shippingTotal}
