+++ src/features/Checkout/components/SuccessStep/style.scss
